Restore incorrectly removed sudo.tmpfiles
Per CLEAR-863, the directory indicated in our configuration is a *mandatory*
one, as #includedir is in fact fatal when the directory does not exist. Due
to this, our sudoers configuration cannot be overriden in a manner consistent
with our stateless policies.

Thus, this package must always provide the sudoers.d directory, empty or not.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>